requirements.txt
django
djangorestframework
psycopg2-binary
flake8
gunicorn

